requirements.txt
# Production dependencies
fastmcp>=2.10.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
snowflake-sqlalchemy>=1.5.0,<2.0.0
snowflake-connector-python>=3.0.0,<4.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
rdflib>=7.0.0,<8.0.0
owlrl>=6.0.0,<7.0.0
